feat(datagrid-number-filter-web): removing logic from rendering component
