+++ packages/pluggableWidgets/datagrid-number-filter-web/src/DatagridNumberFilter.tsx
@@ -4,6 +4,26 @@ import { DatagridNumberFilterContainerProps } from "../typings/DatagridNumberFil
 import { FilterComponent } from "./components/FilterComponent";
 import { getFilterDispatcher } from "./utils/provider";
 import { Alert } from "@mendix/piw-utils-internal";
+import { Big } from "big.js";
+import { DefaultFilterEnum } from "../../datagrid-text-filter-web/typings/DatagridTextFilterProps";
+
+import {
+    attribute as attributeFunction,
+    equals,
+    greaterThan,
+    greaterThanOrEqual,
+    lessThan,
+    lessThanOrEqual,
+    literal,
+    notEqual
+} from "mendix/filters/builders";
+import { FilterCondition } from "mendix/filters";
+
+function getAttributeTypeErrorMessage(type?: string): string | null {
+    return type && !type.match(/AutoNumber|Decimal|Integer|Long/)
+        ? "The attribute type being used for Data grid number filter is not 'Auto number, Decimal, Integer or Long'"
+        : null;
+}
 
 export default function DatagridNumberFilter(props: DatagridNumberFilterContainerProps): ReactElement {
     const FilterContext = getFilterDispatcher();
@@ -14,14 +34,49 @@ export default function DatagridNumberFilter(props: DatagridNumberFilterContaine
     );
     return FilterContext?.Consumer ? (
         <FilterContext.Consumer>
-            {({ filterDispatcher, attribute }) =>
-                filterDispatcher ? (
+            {filterContextValue => {
+                if (!filterContextValue || !filterContextValue.filterDispatcher || !filterContextValue.attribute) {
+                    return alertMessage;
+                }
+                const { filterDispatcher, attribute } = filterContextValue;
+
+                const errorMessage = getAttributeTypeErrorMessage(attribute.type);
+                if (errorMessage) {
+                    return <Alert bootstrapStyle="danger">{errorMessage}</Alert>;
+                }
+
+                const getFilterConditions = (
+                    value: Big | undefined,
+                    type: DefaultFilterEnum
+                ): FilterCondition | undefined => {
+                    if (!attribute || !attribute.filterable || !value) {
+                        return undefined;
+                    }
+                    const filterAttribute = attributeFunction(attribute.id);
+                    const filterValue = new Big(value);
+
+                    switch (type) {
+                        case "greater":
+                            return greaterThan(filterAttribute, literal(filterValue));
+                        case "greaterEqual":
+                            return greaterThanOrEqual(filterAttribute, literal(filterValue));
+                        case "equal":
+                            return equals(filterAttribute, literal(filterValue));
+                        case "notEqual":
+                            return notEqual(filterAttribute, literal(filterValue));
+                        case "smaller":
+                            return lessThan(filterAttribute, literal(filterValue));
+                        case "smallerEqual":
+                            return lessThanOrEqual(filterAttribute, literal(filterValue));
+                    }
+                };
+                return (
                     <FilterComponent
                         adjustable={props.adjustable}
-                        attribute={attribute}
                         defaultFilter={props.defaultFilter}
                         delay={props.delay}
                         filterDispatcher={filterDispatcher}
+                        getFilterConditions={getFilterConditions}
                         name={props.name}
                         placeholder={props.placeholder?.value}
                         screenReaderButtonCaption={props.screenReaderButtonCaption?.value}
@@ -29,10 +84,8 @@ export default function DatagridNumberFilter(props: DatagridNumberFilterContaine
                         tabIndex={props.tabIndex}
                         value={props.defaultValue?.value}
                     />
-                ) : (
-                    alertMessage
-                )
-            }
+                );
+            }}
         </FilterContext.Consumer>
     ) : (
         alertMessage
